Return backwards compatible JSON for ELR status endpoint

+++ data-ingestion-service/src/main/java/gov/cdc/dataingestion/reportstatus/model/ElrStatus.java
@@ -5,5 +5,5 @@ import java.util.UUID;
 
 public record ElrStatus(UUID id, String status, List<Detail> details) {
 
-  public record Detail(long messageId, String status) {}
+  public record Detail(Integer messageId, String status) {}
 }
